+++ src/restfw/testing/__init__.py
@@ -3,11 +3,6 @@
 :Authors: cykooz
 :Date: 06.12.2016
 """
-import base64
-
-from ..utils import force_utf8
-
-
 try:
     import pytest
     pytest.register_assert_rewrite(
@@ -18,8 +13,3 @@ except (ImportError, AttributeError):
     pass
 
 from .resource_testing import assert_resource
-
-
-def basic_auth_value(user_name, password):
-    auth_str = force_utf8(u'%s:%s' % (user_name, password))
-    return 'Basic {}'.format(base64.b64encode(auth_str).decode())